restructure the repo a bit
